+++ src/main/kotlin/io/openfuture/chain/service/Services.kt
@@ -139,7 +139,7 @@ interface StakeholderService {
 
 interface ConsensusService {
 
-    fun getCurrentEpochHeight(): Long
+    fun getCurrentEpochHeight(): Int
 
     fun isGenesisBlockNeeded(): Boolean
